pyproject.toml
[project]
name = "exam-coach"
version = "0.1.0"
description = "Hybrid question-bank and vector-retrieval backend for physics exam generation."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "llama-parse>=0.6.94",
    "pydantic>=2.0.0",
    "pypdf>=5.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["exam_coach*"]
